requirements.txt
annotated-types==0.7.0
Authlib==1.5.2
bandit==1.8.3
boolean.py==5.0
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
cyclonedx-python-lib==8.9.0
defusedxml==0.7.1
dparse==0.6.4
filelock==3.16.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
license-expression==30.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
nltk==3.9.1
packageurl-python==0.16.0
packaging==24.2
pbr==6.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.8.0
platformdirs==4.3.7
psutil==6.1.1
py-serializable==1.1.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resend==2.7.0
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety==3.3.1
safety-schemas==0.0.11
shellingham==1.5.4
sortedcontainers==2.4.0
stevedore==5.4.1
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.13.1
urllib3==2.3.0
